+++ pkg/controller/nodeconfig/sync_rolebindings.go
@@ -8,7 +8,6 @@ import (
 
 	scyllav1alpha1 "github.com/scylladb/scylla-operator/pkg/api/scylla/v1alpha1"
 	"github.com/scylladb/scylla-operator/pkg/controllerhelpers"
-	"github.com/scylladb/scylla-operator/pkg/naming"
 	"github.com/scylladb/scylla-operator/pkg/resourceapply"
 	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -23,9 +22,9 @@ func (ncc *Controller) syncRoleBindings(
 	var progressingConditions []metav1.Condition
 
 	requiredRoleBindings := []*rbacv1.RoleBinding{
-		makePerftuneRoleBinding(),
-		makeSysctlsRoleBinding(),
-		makeRlimitsRoleBinding(),
+		makePerftuneRoleBinding(ncc.namespace),
+		makeSysctlsRoleBinding(ncc.namespace),
+		makeRlimitsRoleBinding(ncc.namespace),
 	}
 
 	// Delete any excessive RoleBindings.
@@ -35,7 +34,7 @@ func (ncc *Controller) syncRoleBindings(
 		requiredRoleBindings,
 		roleBindings,
 		&controllerhelpers.PruneControlFuncs{
-			DeleteFunc: ncc.kubeClient.RbacV1().RoleBindings(naming.ScyllaOperatorNodeTuningNamespace).Delete,
+			DeleteFunc: ncc.kubeClient.RbacV1().RoleBindings(ncc.namespace).Delete,
 		},
 		ncc.eventRecorder)
 	if err != nil {